Update edit and delete comment controller

+++ controllers/comment.controller.js
@@ -2,10 +2,20 @@ const Comment = require("../models/comment.model");
 
 exports.comment_edit = async (req, res) => {
   try {
-    if (!req.body.text) return res.status(404).json({ msg: "text was empty!" });
-    const comment = await Comment.findById(req.params.commentId);
+    if (!req.body.text) return res.status(400).json({ msg: "text was empty!" });
 
+    const comment = await Comment.findById(req.params.commentId).exec();
+
+    // console.log({ comment });
+    // console.log({ commentAuthor: comment.author.toString() });
     comment.text = req.body.text;
+
+    const commentAuthorId = comment.author.toString();
+
+    if (commentAuthorId !== req.userAuth.id)
+      return res
+        .status(400)
+        .json({ message: "You cannot remove other user comment!" });
 
     await comment.save();
 
@@ -18,6 +28,13 @@ exports.comment_edit = async (req, res) => {
 exports.comment_delete = async (req, res) => {
   try {
     const comment = await Comment.findById(req.params.commentId);
+
+    const commentAuthorId = comment.author.toString();
+
+    if (commentAuthorId !== req.userAuth.id)
+      return res
+        .status(400)
+        .json({ message: "You cannot remove other user comment!" });
 
     await comment.deleteOne();
